# Avoid making extra 5 arrays in each `save`